backend-py/requirements-common.txt
# Common requirements used across local dev and production
FastAPI==0.111.0
gunicorn==20.0.4
starlette==0.37.2
uvicorn==0.25.0
google-cloud-ndb
passlib==1.7.4
python-jose==3.3.0 
black
faker=25.2.0